% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{estimate_pi2}
\alias{estimate_pi2}
\title{Estimate Pi from random sampling with Rcpp integration}
\usage{
estimate_pi2(B = 5000, seed = 10)
}
\arguments{
\item{B}{A \code{numeric} (integer) used to denote the number of sample
coordinates used to approximate pi}

\item{seed}{A \code{numeric} used to control the seed of the random number
generator used by this function.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{estimated_pi}{Estimated value of pi}
     \item{points}{Dataframe of the generated coordinates with an indication
     whether or not the coordinate is inside the circle}
}
}
\description{
Compute an approximation of pi by drawing x and y coordinates
from a uniform distribution and checking whether or not they fall inside a
unit circle.
}
\examples{
estimate_pi2(B=10)
}
\author{
Vanessa Armistead

Bart Roes

Luisa Pricken

Ameni Rouatbi
}
