% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{plot.pi}
\alias{plot.pi}
\title{Plot objects of class pi}
\usage{
\method{plot}{pi}(x)
}
\arguments{
\item{x}{A \code{list} of type pi}
}
\value{
A \code{plot} of the object
}
\description{
Creating a plot representing the random coordinates created by
the function "estimate_pi".
}
\examples{
obj <- estimate_pi()
plot(obj)
}
\author{
Vanessa Armistead

Bart Roes

Luisa Pricken

Ameni Rouatbi
}
