% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemo.R
\name{runDemo}
\alias{runDemo}
\title{Run Shiny app}
\usage{
runDemo()
}
\description{
This function allows for launching the Shiny app when the
package is loaded.
This app will allow you to compare the two different functions that estimate
pi.
}
\author{
Vanessa Armistead

Bart Roes

Luisa Pricken

Ameni Rouatbi
}
